+++ tests/views.py
@@ -9,11 +9,16 @@ from django.template.response import TemplateResponse
 from django.views.decorators.cache import cache_page
 
 from debug_toolbar.utils import get_csp_nonce
-from tests.models import PostgresJSON
+from tests.models import Binary, PostgresJSON
 
 
 def execute_sql(request):
     list(User.objects.all())
+    return render(request, "base.html")
+
+
+def execute_binary_sql(request):
+    list(Binary.objects.filter(field=b"\x01\x02\x03"))
     return render(request, "base.html")